feat: adding web listener, use TCP instead
