--- proto/crawling.proto
@@ -1,26 +1,0 @@
-syntax = "proto3";
-package crawling;
-
-service Crawler {
-  rpc CrawlUrl (CrawlingRequest) returns (CrawlingReply);
-}
-
-message CrawlingRequest {
-  string host = 1;
-  string uri = 2;
-  int64  depth = 3;
-}
-
-message CrawlingReply {
-  Frontier frontier = 1;
-}
-
-message Frontier {
-  repeated Extract extracts = 1;
-}
-
-message Extract {
-  string domain = 1;
-  string uri = 2;
-}
-
