+++ src/crawler/distributed.rs
@@ -1,0 +1,6 @@
+pub fn execute(
+    start_url: String,
+    web_url: String
+) -> Result<Vec<(String, String)>, Box<dyn std::error::Error + Send + Sync>> {
+    let mut
+}
