--- src/grpc/mod.rs
@@ -1,2 +1,0 @@
-pub mod client;
-pub mod server;
