+++ src/main.rs
@@ -6,11 +6,11 @@ mod argparser;
 mod crawler;
 mod errors;
 mod frontier;
-mod grpc;
 mod lib;
 mod logger;
 mod net;
 mod url;
+mod web;
 
 use argparser::Program;
